Fixed vitest templates
'registerNodeLoader' on true wouldn't let the tests run successfully, not in this version, nor in newer. It also is necessary to have it explicitly on false, or else, the tests wouldn't work either.

Besides it doesn't log any warning on dependency duplicity, like it does in other cases, so I also removed the comment that said that.